Test parfor with full problem size
Tests using previous commit were successful for R2014a, R2015a and R2016a.
This is the original (full) scale problem which was sometimes crashing
for the user.
I have only tested it so far using R2015a, so will repeat using R2014a,
which is what the user is using.